Workaround for dbuild's scalaBinaryVersion mishandling.
This is another workaround for typesafehub/dbuild#145. This time we deal
with the problem of Scala version-dependent inclusion of source files to
be compiled. Selection of a source folder is broken because it is based
on scalaBinaryVersion which dbuild overrides. The workaround is to
fallback to "2.11" value for scalaBinaryVersion when the one set in the
build is not included in the known set ("2.9.3, "2.10", "2.11").