Change template import priorities
Imports are now added in different scopes so they can have different
priorities. The lowest priority are the imports included with every
template. Next in priority are the imports provided through the sbt
plugin setting. Highest in priority are imports included in the
template itself.

This fixes a regression in import behaviour from Play 2.3 to Play
2.4. However, there is one change. Play 2.3 had two levels of
import priority, but this change introduces three levels, which
should more often be what users want.

The code generation has changed slightly, with outer objects being
generated if necessary and the template code being generated within
a class in those objects. The static template object extends that
class.

For efficiency, the code generation takes account of whether or not
there are any extra imports and only generates extra scope objects
if they are needed.