Make docs dependency on api an inter project dependency
Although the twirl sbt plugin automatically adds a dependency on the
correct version of the api dependency, that requires the dependency to
be published first. This fix makes the dependency an inter project
dependency, which fixes a problem in our travis build where if the API
is changed, the docs build, which is a separate build from the main
build, doesn't see the changes, for example as seen in #86.