Injectable templates
Fixes #93
Fixes #71

Adds support for injectable templates.

The syntax for this is:

    @this(...)

Multiple parameter lists and implicit parameters are supported.  The
last comment before the constructor is used as the comment for the
constructor in the generated scala code.

In addition to implementing this, I greatly improved the way start of
the file is passed - you can now put comments between import statements,
and as many comments as you want between the top import statements and
the parameter declaration (previously, you could only have one comment
there).

There is one breaking change that I had to make to do this - previously
imports were scoped using a nested object pattern, allowing imports in
the templates to override the imports defined in sbt, and those imports
to override the default imports.  This feature has been removed, all
these imports are in the same scope.  To compensate, I've done the
following:

* There are now no default hard coded imports, all imports are configured
  in sbt. This means they can be completely overridden.  This alleviates
  the need to have the default imports in a lower priority scope to the
  sbt imports.
* I've greatly limited what gets imported, there is no longer a blanket
  import on play.twirl.api._.  Only things that are needed are imported.
* I've deprecated TemplateMagic (it's not magic, bad name), and it no
  longer gets imported.  It's kept around for binary compatibility (so
  templates compiled with the old Twirl will still work).
* I've replaced TemplateMagic with two objects, TwirlFeatureImports, for
  features of the Twirl language, and TwirlHelperImports, for helpers.
  In addition, all the symbols exposed by importing these two have been
  namespaced, so there should be no naming collisions, making it less
  necessary to scope imports.
* In general, going forward, we should discourage using imports defined
  in sbt, and limit what Play adds, since it generally introduces more
  confusion than not (why do I need to import this, but not something in
  the models package?)  Using Twirl should be closer to regular Scala in
  its rules and imports, the only things automatically imported should
  be things that are generally considered features of the Twirl
  language.