Added security info to template escaping (#183)
* Added security info to template escaping

A very brief explanation of template escaping is given (with code sample), but there is currently no warning of the security implications of using this technique. Given the potential for harm, I think it would be helpful for developers to understand the risks in using this.

I've added a short description as a starting point for discussion and would be grateful to receive feedback on it. I considered including links to more detailed content on XSS but wasn't sure if that fits with the documentation guidelines for twirl. Happy to add some links if anyone thinks that would be beneficial.

* Moved new text into a 'Note' as per suggestion