Travis-CI: use standardized JDK+sbt setup
The issue here is that by default, Travis-CI has an outdated
sbt-extras script that doesn't work with sbt 1.4.8+.

A more minimal fix would be to add the workaround suggested
in the sbt 1.4.8 and 1.4.9 release notes.

But down that path lies further hand-maintenance.  I think it's better
to use the Travis-CI stuff in the scala-dev repo, which is maintained
by the Scala team at lightbend. It handles JDK installation (via
SDKMAN not Jabba), addresses the sbt 1.4.8+ issue (and will address
any other such issues that arise in the future), and handles
caching, too.