CI with GitHub Actions
Replaced the existing Travis configuration with the GitHub Actions configuration.
I used the play-json configuration as a reference, especially for publish.yml.
I've confirmed that the validation and testing work, but I haven't been able to confirm that the release works.

Notifications to Gitter have not been migrated. (I wonder if this Gitter channel is still active?)